+++ src/dbQueries/users.ts
@@ -1,8 +1,8 @@
 import { createUserToken, creationDates } from "../helperFunctions";
-import { isCreateUserBody } from "../validationFunctions";
 import bcrypt from "bcrypt";
 import { ReqBuilder } from "../middleware/auth_types";
 import { LoginBody } from "./request_types";
+import { isCreateUserBody, isLoginBody } from "../json_validation/request_body";
 
 // Create
 export const createUser: ReqBuilder =
@@ -34,6 +34,10 @@ export const getAllUsers: ReqBuilder = (client) => async (req, res) => {
 };
 
 export const loginUser: ReqBuilder = (client) => async (req, res) => {
+  if (!isLoginBody(req.body)) {
+    res.status(400).json({ message: "Invalid email or password" });
+    return;
+  }
   const { email, password } = req.body as LoginBody;
   const user = await client.user.findFirst({
     where: {
